fix: change js to jsx and add header navlink
