+++ package.json
@@ -8,6 +8,7 @@
         "@testing-library/user-event": "^13.5.0",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
+        "react-router-dom": "^6.23.1",
         "react-scripts": "5.0.1",
         "styled-component": "^2.8.0",
         "web-vitals": "^2.1.4"
