--- src/components/App.js
@@ -1,14 +1,0 @@
-import Header from './Header';
-import '../styles/fonts.css';
-import '../styles/color.css';
-import '../styles/reset.css';
-
-function App() {
-    return (
-        <>
-            <Header />
-        </>
-    );
-}
-
-export default App;
