+++ src/components/Items.jsx
@@ -1,0 +1,14 @@
+// import BestProductList from './BestProductList';
+// import ProductList from './ProductList';
+
+function Main() {
+    return (
+        <>
+            <h2> 테스트용</h2>
+            {/* <BestProductList />
+            <ProductList /> */}
+        </>
+    );
+}
+
+export default Main;
