--- src/components/Main.js
@@ -1,13 +1,0 @@
-import BestProductList from './BestProductList';
-import ProductList from './ProductList';
-
-function Main() {
-    return (
-        <>
-            <BestProductList />
-            <ProductList />
-        </>
-    );
-}
-
-export default Main;
