+++ src/styles/NavigationLink.module.css
@@ -18,10 +18,6 @@
     color: var(--gray600);
 }
 
-.nav:hover {
-    color: var(--blue);
-}
-
 .login {
     display: flex;
     align-items: center;
@@ -37,6 +33,10 @@
 
 .login:hover {
     background: #1967d6;
+}
+
+a.active {
+    color: var(--blue);
 }
 
 @media (min-width: 768px) and (max-width: 1199px) {
